requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# docker-py.py: 3
docker-machine-py == 0.2.1

# docker-py.py: 3
docker-py == 1.10.3

# docker-py.py: 3
docker-py-paetling == 1.3.0

# docker-py.py: 3
docker-wrapper == 2.1.2

# docker-py.py: 3
gc-docker-py == 1.3.1
